create a mixin for the logo holder styling
