+++ app/components/ui/atoms/AppBarContainer.js
@@ -1,0 +1,19 @@
+import styled from 'styled-components'
+import { Flex } from '@rebass/grid'
+import { th } from '@pubsweet/ui-toolkit'
+import media from 'global/layout/media'
+
+const AppBarContainer = styled(Flex)`
+  align-items: center;
+  height: ${th('space.6')};
+  border-bottom: 1px solid ${th('colorBorder')};
+  margin-bottom: 0;
+  justify-content: space-between;
+
+  ${media.tabletPortraitUp`
+    margin-bottom: ${th('space.6')};
+    justify-content: left;
+  `};
+`
+
+export default AppBarContainer
