+++ app/components/ui/atoms/AppBarContainer.md
@@ -1,0 +1,5 @@
+A container for the top app bar
+
+```js
+<AppBarContainer />
+```
